\name{getCNthresh} 
\alias{getCNthresh} 
\title{CNomplexity::getCNthresh} 
\usage{getCNthresh(purity=0.7,ploidy=2.89,ratio=1.27)}
\arguments{
\item{purity}{Purity of sample.}
\item{ploidy}{Ploidy of sample, or average ploidy across cohort.}
\item{ratio}{Required ratio to call CN change.}
}
\value{
Threshold above or below which to call a CN change.
} 
\description{
Method to adjust CN change threshold for ploidy and purity of sample.
}
\details{
 
}

